backend/requirements.txt
# SnapLinked v3.0 - Dependências Python Seguras
Flask==3.0.3
Flask-CORS==5.0.0
Werkzeug==3.0.4
python-dotenv==1.0.1
playwright==1.48.0
SQLAlchemy==2.0.35
Flask-SQLAlchemy==3.1.1
bcrypt==4.2.0
PyJWT==2.9.0
requests==2.32.3
cryptography==43.0.1
